+++ MainServer/Api/__init__.py
@@ -6,6 +6,7 @@ from ..Api.teams import router as team_router
 from ..Api.contests import router as contest_router
 from ..Api.tasks import router as task_router
 from ..Api.user_contest_view import router as user_contest_view_router
+from ..Api.compilations import router as compilations_router
 
 
 router = APIRouter()
@@ -16,3 +17,4 @@ router.include_router(team_router)
 router.include_router(contest_router)
 router.include_router(task_router)
 router.include_router(user_contest_view_router)
+router.include_router(compilations_router)
