+++ MainServer/Models/Compilations.py
@@ -1,0 +1,13 @@
+from typing import List
+from pydantic import BaseModel
+
+
+class BaseCompilations(BaseModel):
+    name: str
+    extend: str
+
+
+class GetCompilations(BaseCompilations):
+    id: int
+    class Config:
+           orm_mode = True
