+++ MainServer/async_database.py
@@ -1,0 +1,20 @@
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.ext.asyncio import create_async_engine
+from sqlalchemy.orm import sessionmaker
+from .settings import settings
+
+
+DATABASE_URL = f"postgresql+asyncpg://{settings.database_name.strip()}"
+
+
+engine = create_async_engine(DATABASE_URL, echo=True)
+Base = declarative_base()
+async_session = sessionmaker(
+    engine, class_=AsyncSession, expire_on_commit=False
+)
+
+
+async def get_session() -> AsyncSession:
+    async with async_session() as session:
+        yield session
