+++ settings_server.env
@@ -7,3 +7,6 @@ SERVER_REVIEW_PORT=50001
 DATABASE_NAME=admin:admin@localhost:5432/contest
 STATIC_PATH=Files
 JWT_SECRET=HvHmOQ9_7cliZgsVSFeaGE8723pZUrxlDfDRL-NlW6Q
+
+FRONT_END_HOST=localhost
+FRONT_END_PORT=8080
